feat: 其他设备及出入库模块接口开发

+++ src/main/java/cn/shmedo/monitor/monibotbaseapi/model/param/project/AddProjectParam.java
@@ -118,10 +118,10 @@ public class AddProjectParam implements ParameterValidator, ResourcePermissionPr
                 return ResultWrapper.withCode(ResultCode.INVALID_PARAMETER, "模板与项目不适配");
             }
         }
-        if (!PredefinedModelProperTyCache.projectTypeAndPropertyListMap.containsKey(projectType)) {
+        if (!PredefinedModelProperTyCache.projectTypeAndPropertyListMap.containsKey(projectType.intValue())) {
             return ResultWrapper.withCode(ResultCode.INVALID_PARAMETER, "系统无此项目类型的预定义模板");
         }
-        properties = new ArrayList<>(PredefinedModelProperTyCache.projectTypeAndPropertyListMap.get(projectType));
+        properties = new ArrayList<>(PredefinedModelProperTyCache.projectTypeAndPropertyListMap.get(projectType.intValue()));
 
         if (modelID != null) {
             TbPropertyMapper tbPropertyMapper = ContextHolder.getBean(TbPropertyMapper.class);
